requirements.txt
streamlit
pandas
langchain
langchain-huggingface
huggingface_hub
python-dotenv
